+++ lib/td_dd_web/controllers/data_structure_version_controller.ex
@@ -29,7 +29,6 @@ defmodule TdDdWeb.DataStructureVersionController do
     :data_field_links,
     :versions,
     :system,
-    :ancestry,
     :links,
     :profile,
     :degree,