+++ lib/td_dd_web/controllers/metadata_controller.ex
@@ -77,7 +77,7 @@ defmodule TdDdWeb.MetadataController do
              external_id: external_id,
              parent_external_id: parent_external_id
            ),
-         dsv <- DataStructures.get_latest_version_by_external_id(external_id, enrich: [:ancestry]) do
+         dsv <- DataStructures.get_latest_version_by_external_id(external_id) do
       render(conn, "show.json", data_structure_version: dsv)
     else
       {:can, false} -> {:can, false}
@@ -90,7 +90,7 @@ defmodule TdDdWeb.MetadataController do
 
     with {:can, true} <- {:can, can_upload?(user, params)},
          {:ok, _} <- do_upload(conn, params, external_id: external_id),
-         dsv <- DataStructures.get_latest_version_by_external_id(external_id, enrich: [:ancestry]) do
+         dsv <- DataStructures.get_latest_version_by_external_id(external_id) do
       render(conn, "show.json", data_structure_version: dsv)
     end
   end