+++ test/td_dd/loader/structures_test.exs
@@ -1,8 +1,8 @@
-defmodule TdDd.Loader.DomainIdLoader.Test do
+defmodule TdDd.Loader.StructuresTest do
   use TdDd.DataCase
 
   alias TdDd.DataStructures.DataStructure
-  alias TdDd.Loader.DomainIdLoader
+  alias TdDd.Loader.Structures
 
   setup %{ids: ids} do
     %{id: system_id} = insert(:system)
@@ -30,19 +30,19 @@ defmodule TdDd.Loader.DomainIdLoader.Test do
     test "bulk-updates domain_id iff changed, returns count and affected ids", %{ids: ids} do
       ts = DateTime.utc_now() |> DateTime.truncate(:second)
 
-      assert {0, []} = DomainIdLoader.bulk_update_domain_id([], 1, ts)
+      assert {0, []} = Structures.bulk_update_domain_id([], 1, ts)
 
       assert {2_500, updated_ids} =
                ids
                |> Enum.map(&Integer.to_string/1)
-               |> DomainIdLoader.bulk_update_domain_id(1, ts)
+               |> Structures.bulk_update_domain_id(1, ts)
 
       assert length(updated_ids) == 2_500
 
       assert {0, []} =
                updated_ids
                |> Enum.map(&Integer.to_string/1)
-               |> DomainIdLoader.bulk_update_domain_id(1, ts)
+               |> Structures.bulk_update_domain_id(1, ts)
     end
 
     @tag ids: 1..10
@@ -50,8 +50,8 @@ defmodule TdDd.Loader.DomainIdLoader.Test do
       ts = DateTime.utc_now() |> DateTime.truncate(:second)
       external_ids = Enum.map(ids, &Integer.to_string/1)
 
-      assert {10, _} = DomainIdLoader.bulk_update_domain_id(external_ids, nil, ts)
-      assert {0, []} = DomainIdLoader.bulk_update_domain_id(external_ids, nil, ts)
+      assert {10, _} = Structures.bulk_update_domain_id(external_ids, nil, ts)
+      assert {0, []} = Structures.bulk_update_domain_id(external_ids, nil, ts)
     end
   end
 end